+++ include/variables.h
@@ -195,7 +195,7 @@ extern Color_RGBA8_u32 D_801614B0;
 extern PreNmiBuff* gAppNmiBufferPtr;
 extern SchedContext gSchedContext;
 extern PadMgr gPadMgr;
-extern u32 gSegments[NUM_SEGMENTS];
+extern uintptr_t gSegments[NUM_SEGMENTS];
 extern volatile OSTime D_8016A520;
 extern volatile OSTime D_8016A528;
 extern volatile OSTime D_8016A530;
